Add BatchParam into its own file rather than batcher.go
